Refactor InventoryMenu to use generic Menu class
