+++ src/Rpg/mainCharacter/MainCharacter.cpp
@@ -23,10 +23,12 @@ MainCharacter::MainCharacter(const sf::Vector2f& position, Inventory* inventory,
     mAnimations[int(AnimationIndex::WalkingLeft)] = Animation(0, 0, 64, 64, "assets/sprites/mainCharacter/walkLeft.png", 6, 1.52f);
     mAnimations[int(AnimationIndex::WalkingRight)] = Animation(0, 0, 64, 64, "assets/sprites/mainCharacter/walkRight.png", 6, 1.52f);
 
-    mIconSprite.setTextureRect({ 0, 0, 64, 64 });
-    mIconTexture = Animation(0, 0, 28, 64, "assets/sprites/mainCharacter/idleDown.png", 1, 0.0f);
-    mIconTexture.applyToSprite(mIconSprite);
-    mIconSprite.setScale({ 2.7f, 2.7f });
+    if (mAvatarTexture.loadFromFile("assets/sprites/mainCharacter/playerAvatar.png")) {
+        mAvatarSprite.setTexture(mAvatarTexture);
+
+        mAvatarSprite.setOrigin(mAvatarSprite.getLocalBounds().width / 2.f, mAvatarSprite.getLocalBounds().height / 2.f);
+        mAvatarSprite.setScale(3.f, 3.f);
+    }
 }
 
 void MainCharacter::update(float dt, bool inDialogue) {
@@ -206,6 +208,6 @@ sf::Sprite& MainCharacter::getSprite() {
     return mSprite;
 }
 
-sf::Sprite& MainCharacter::getIconSprite() {
-    return mIconSprite;
+sf::Sprite& MainCharacter::getAvatarSprite() {
+    return mAvatarSprite;
 }
