+++ src/Rpg/menues/InventoryMenu.h
@@ -1,22 +1,21 @@
 #pragma once
 #include <SFML/Graphics.hpp>
+#include "../../ui/Menu.h"
 #include "../mainCharacter/MainCharacter.h"
-#include "Inventory.h"
-#include "items/DroppedItem.h"
+#include "../inventory/Inventory.h"
+#include "../inventory/items/DroppedItem.h"
 
-class InventoryMenu {
+class InventoryMenu : public Menu {
 public:
-    InventoryMenu(Inventory& inventory, const sf::Vector2f& position,
+    InventoryMenu(const sf::Vector2f& position, const sf::Vector2f& size, Inventory& inventory,
                   MainCharacter& character, const sf::Vector2f& slotSize);
 
-    void render(sf::RenderWindow& window);
-    void handleMouseClick(const sf::Vector2f& mousePos);
-    void updateHover(const sf::Vector2f& mousePos);
-    void handleDragAndDrop(const sf::Vector2f& mousePos);
+    void render(sf::RenderWindow& window) override;
+    void handleMouseClick(const sf::Vector2f& mousePos) override;
+    void updateHover(const sf::Vector2f& mousePos) override;
+    void update(float dt) override;
 
-    void update();
     int getHoveredSlot() const;
-
     void restart();
 
 private:
@@ -26,7 +25,7 @@ private:
     Inventory& mInventory;
     std::vector<sf::RectangleShape> mSlots;
     sf::Vector2f mSlotSize;
-    sf::Vector2f mPosition;
+
     int mHoveredSlot;
     int mDraggedSlot;
     float mGap;
@@ -39,10 +38,11 @@ private:
     // Tooltip elements
     sf::RectangleShape mTooltipBackground;
     sf::Text mTooltipText;
+    bool mShowTooltip;
     sf::Font mFont;
 
+    void initializeLayout();
     void updateSlotColors();
     int getSlotIndexAtPosition(const sf::Vector2f& pos) const;
     void updateTooltip();
 };
-
